refactor: extract runner character manifest support
