+++ Makefile
@@ -5,6 +5,7 @@ check:
 	python -m py_compile \
 		app/services/runner.py \
 		app/services/runner_audio_render_support.py \
+		app/services/runner_character_manifest.py \
 		app/services/runner_image_review.py \
 		app/services/runner_render_plan.py \
 		app/services/runner_session.py \
