+++ app/services/runner.py
@@ -22,6 +22,7 @@ from app.schemas.workflow import (
     WorkflowRunResponse,
 )
 from app.services.runner_audio_render_support import RunnerAudioRenderSupport
+from app.services.runner_character_manifest import RunnerCharacterManifestSupport
 from app.services.runner_errors import UnknownStepError, UnknownVideoProviderError
 from app.services.runner_image_review import RunnerImageReviewSupport
 from app.services.runner_image_selection_support import RunnerImageSelectionSupport
@@ -149,6 +150,7 @@ class WorkflowRunner:
 
     def __init__(self) -> None:
         self._audio_render_support = RunnerAudioRenderSupport(self)
+        self._character_manifest_support = RunnerCharacterManifestSupport(self)
         self._image_selection_support = RunnerImageSelectionSupport(self)
         self._image_review = RunnerImageReviewSupport(self)
         self._render_plan = RunnerRenderPlanSupport(self)
@@ -359,7 +361,9 @@ class WorkflowRunner:
         audience_label = self._audience_label(ctx.input.audience)
         story_plan = self._duration_story_plan(ctx.input.duration_sec)
 
-        manifest_items = self._character_manifest_items(outputs)
+        manifest_items = self._character_manifest_support.character_manifest_items(
+            outputs
+        )
         character_hint = ""
         if manifest_items:
             parts = []
@@ -731,8 +735,10 @@ class WorkflowRunner:
 
         step_results: List[StepResult] = []
 
-        character_candidates = self._build_character_candidates(req.input)
-        character_manifest = self._build_character_manifest(
+        character_candidates = self._character_manifest_support.build_character_candidates(
+            req.input
+        )
+        character_manifest = self._character_manifest_support.build_character_manifest(
             req.input, character_candidates
         )
 
@@ -775,7 +781,7 @@ class WorkflowRunner:
                 if inferred_primary is not None:
                     character_manifest = [inferred_primary]
 
-        character_manifest = self._enrich_character_manifest_traits_from_topic(
+        character_manifest = self._character_manifest_support.enrich_character_manifest_traits_from_topic(
             character_manifest,
             req.input.topic,
         )
@@ -867,303 +873,6 @@ class WorkflowRunner:
             timestamp=WorkflowRunResponse.now_timestamp(),
         )
 
-    def _build_character_candidates(
-        self, workflow_input: WorkflowInput
-    ) -> List[Dict[str, Any]]:
-        if workflow_input.structured_characters_enabled and workflow_input.characters:
-            items: List[Dict[str, Any]] = []
-            for index, character in enumerate(workflow_input.characters, start=1):
-                items.append(
-                    {
-                        "candidate_id": f"candidate_{index:02d}",
-                        "display_name": character.display_name.strip(),
-                        "species": character.species.strip(),
-                        "role_type": character.role_type,
-                        "visual_traits": character.visual_traits.strip(),
-                        "forbidden_traits": character.forbidden_traits.strip(),
-                        "source": "structured_input",
-                    }
-                )
-            return items
-
-        items: List[Dict[str, Any]] = []
-
-        main_display_name = (
-            workflow_input.main_character_display.strip()
-            or workflow_input.main_character.strip()
-        )
-        if main_display_name:
-            items.append(
-                {
-                    "candidate_id": "candidate_01",
-                    "display_name": main_display_name,
-                    "species": workflow_input.main_character_species.strip() or "None",
-                    "role_type": "primary",
-                    "visual_traits": workflow_input.main_character_visual_traits.strip(),
-                    "forbidden_traits": "",
-                    "source": "legacy_main_character",
-                }
-            )
-
-        secondary_display_name = (
-            workflow_input.secondary_character_display.strip()
-            or workflow_input.secondary_character.strip()
-        )
-        if secondary_display_name:
-            items.append(
-                {
-                    "candidate_id": f"candidate_{len(items) + 1:02d}",
-                    "display_name": secondary_display_name,
-                    "species": workflow_input.secondary_character_species.strip()
-                    or "None",
-                    "role_type": "secondary",
-                    "visual_traits": workflow_input.secondary_character_visual_traits.strip(),
-                    "forbidden_traits": "",
-                    "source": "legacy_secondary_character",
-                }
-            )
-
-        return items
-
-    def _build_character_manifest(
-        self,
-        workflow_input: WorkflowInput,
-        candidates: List[Dict[str, Any]],
-    ) -> List[Dict[str, Any]]:
-        manifest: List[Dict[str, Any]] = []
-
-        for index, candidate in enumerate(candidates, start=1):
-            role_type = str(candidate.get("role_type") or "secondary").strip()
-            if role_type == "primary":
-                character_id = "char_primary_01"
-            else:
-                secondary_index = (
-                    sum(
-                        1
-                        for item in manifest
-                        if str(item.get("role_type") or "").strip() == "secondary"
-                    )
-                    + 1
-                )
-                character_id = f"char_secondary_{secondary_index:02d}"
-
-            visual_traits_text = str(candidate.get("visual_traits") or "").strip()
-            forbidden_traits_text = str(candidate.get("forbidden_traits") or "").strip()
-
-            signature_traits = [
-                item.strip() for item in visual_traits_text.split(",") if item.strip()
-            ]
-            forbidden_traits = [
-                item.strip()
-                for item in forbidden_traits_text.split(",")
-                if item.strip()
-            ]
-
-            manifest.append(
-                {
-                    "character_id": character_id,
-                    "display_name": str(candidate.get("display_name") or "").strip(),
-                    "species": str(candidate.get("species") or "").strip(),
-                    "role_type": role_type,
-                    "signature_traits": signature_traits,
-                    "forbidden_traits": forbidden_traits,
-                    "locking_level": "strict",
-                    "reference_assets": {
-                        "status": "pending",
-                        "front_view": None,
-                        "side_view": None,
-                        "three_quarter_view": None,
-                    },
-                    "source": str(candidate.get("source") or "").strip(),
-                }
-            )
-
-        return manifest
-
-    def _extract_color_traits_from_topic(self, topic: str) -> List[str]:
-        normalized = str(topic or "").strip().lower()
-        if not normalized:
-            return []
-
-        color_aliases = [
-            ("red", ["红色", "红", "red"]),
-            ("blue", ["蓝色", "蓝", "blue"]),
-            ("green", ["绿色", "绿", "green"]),
-            ("yellow", ["黄色", "黄", "yellow"]),
-            ("orange", ["橙色", "橙", "orange"]),
-            ("pink", ["粉色", "粉", "pink"]),
-            ("purple", ["紫色", "紫", "purple"]),
-            ("white", ["白色", "白", "white"]),
-            ("black", ["黑色", "黑", "black"]),
-            ("gray", ["灰色", "灰", "grey", "gray"]),
-            ("brown", ["棕色", "棕", "咖啡色", "brown"]),
-        ]
-
-        traits: List[str] = []
-        for english_color, aliases in color_aliases:
-            if any(alias.lower() in normalized for alias in aliases):
-                for alias in aliases:
-                    if alias not in traits:
-                        traits.append(alias)
-                if english_color not in traits:
-                    traits.append(english_color)
-
-        return traits
-
-    def _enrich_character_manifest_traits_from_topic(
-        self,
-        character_manifest: List[Dict[str, Any]],
-        topic: str,
-    ) -> List[Dict[str, Any]]:
-        color_traits = self._extract_color_traits_from_topic(topic)
-        if not color_traits:
-            return character_manifest
-
-        enriched_manifest: List[Dict[str, Any]] = []
-        for item in character_manifest:
-            if not isinstance(item, dict):
-                continue
-
-            enriched_item = dict(item)
-            signature_traits = list(enriched_item.get("signature_traits") or [])
-
-            for trait in color_traits:
-                if trait and trait not in signature_traits:
-                    signature_traits.append(trait)
-
-            enriched_item["signature_traits"] = signature_traits
-            enriched_manifest.append(enriched_item)
-
-        return enriched_manifest
-
-    def _apply_visual_profiles_to_character_manifest(
-        self,
-        character_manifest: Dict[str, Any],
-        character_visual_profiles: Dict[str, Any],
-    ) -> Dict[str, Any]:
-        characters = character_manifest.get("characters") or []
-        profiles = character_visual_profiles.get("profiles") or []
-
-        if not isinstance(characters, list) or not isinstance(profiles, list):
-            return character_manifest
-
-        if not characters or not profiles:
-            return character_manifest
-
-        def normalize(value: Any) -> str:
-            return str(value or "").strip()
-
-        def as_list(value: Any) -> List[str]:
-            if isinstance(value, list):
-                return [normalize(item) for item in value if normalize(item)]
-            if isinstance(value, str):
-                return [
-                    normalize(item)
-                    for item in value.replace("，", ",").replace("；", ",").split(",")
-                    if normalize(item)
-                ]
-            return []
-
-        def extend_unique(base: Any, extra: Any) -> List[str]:
-            result: List[str] = []
-            seen = set()
-
-            for value in as_list(base) + as_list(extra):
-                if value and value not in seen:
-                    result.append(value)
-                    seen.add(value)
-
-            return result
-
-        profile_by_id: Dict[str, Dict[str, Any]] = {}
-        profile_by_name: Dict[str, Dict[str, Any]] = {}
-
-        for profile in profiles:
-            if not isinstance(profile, dict):
-                continue
-
-            character_id = normalize(profile.get("character_id"))
-            display_name = normalize(profile.get("display_name"))
-            species = normalize(profile.get("species"))
-            subject = normalize(profile.get("subject"))
-
-            if character_id:
-                profile_by_id[character_id] = profile
-
-            for key in [display_name, species, subject]:
-                if key:
-                    profile_by_name[key] = profile
-
-        enriched_characters: List[Dict[str, Any]] = []
-
-        for item in characters:
-            if not isinstance(item, dict):
-                continue
-
-            character_id = normalize(item.get("character_id"))
-            display_name = normalize(item.get("display_name"))
-            species = normalize(item.get("species"))
-
-            profile = (
-                profile_by_id.get(character_id)
-                or profile_by_name.get(display_name)
-                or profile_by_name.get(species)
-            )
-
-            if not profile:
-                enriched_characters.append(dict(item))
-                continue
-
-            enriched_item = dict(item)
-
-            enriched_item["signature_traits"] = extend_unique(
-                enriched_item.get("signature_traits") or [],
-                profile.get("must_keep") or [],
-            )
-            enriched_item["forbidden_traits"] = extend_unique(
-                enriched_item.get("forbidden_traits") or [],
-                profile.get("must_avoid") or [],
-            )
-
-            for key in [
-                "visual_identity",
-                "profile_source",
-                "profile_generation_source",
-                "required_presence_rules",
-                "llm_profile_ready",
-                "llm_confidence",
-            ]:
-                value = profile.get(key)
-                if value not in (None, "", []):
-                    enriched_item[key] = value
-
-            enriched_characters.append(enriched_item)
-
-        return {
-            **character_manifest,
-            "count": len(enriched_characters),
-            "characters": enriched_characters,
-        }
-
-    def _character_manifest_items(
-        self, outputs: Dict[str, Any]
-    ) -> List[Dict[str, Any]]:
-        manifest_output = outputs.get("character_manifest") or {}
-        characters = manifest_output.get("characters") or []
-        if isinstance(characters, list):
-            return characters
-        return []
-
-    def _manifest_character_by_role(
-        self,
-        outputs: Dict[str, Any],
-        role_type: str,
-    ) -> Optional[Dict[str, Any]]:
-        for item in self._character_manifest_items(outputs):
-            if str(item.get("role_type") or "").strip() == role_type:
-                return item
-        return None
-
     def _run_image_prompts(
         self, ctx: StepContext, outputs: Dict[str, Any]
     ) -> Dict[str, Any]:
@@ -1199,7 +908,7 @@ class WorkflowRunner:
             outputs,
         )
 
-        enriched_character_manifest = self._apply_visual_profiles_to_character_manifest(
+        enriched_character_manifest = self._character_manifest_support.apply_visual_profiles_to_character_manifest(
             outputs.get("character_manifest") or {},
             character_visual_profiles,
         )
@@ -1412,7 +1121,9 @@ class WorkflowRunner:
         if not isinstance(scene_characters, list):
             return []
 
-        manifest_items = self._character_manifest_items(outputs)
+        manifest_items = self._character_manifest_support.character_manifest_items(
+            outputs
+        )
         manifest_by_id: Dict[str, Dict[str, Any]] = {}
         manifest_by_name: Dict[str, Dict[str, Any]] = {}
 
@@ -1462,7 +1173,7 @@ class WorkflowRunner:
         self, outputs: Dict[str, Any]
     ) -> List[Dict[str, Any]]:
         bindings: List[Dict[str, Any]] = []
-        for item in self._character_manifest_items(outputs):
+        for item in self._character_manifest_support.character_manifest_items(outputs):
             bindings.append(
                 {
                     "character_id": item.get("character_id"),
@@ -1482,7 +1193,7 @@ class WorkflowRunner:
         if not target:
             return None
 
-        for item in self._character_manifest_items(outputs):
+        for item in self._character_manifest_support.character_manifest_items(outputs):
             if str(item.get("character_id") or "").strip() == target:
                 return item
         return None
@@ -2178,7 +1889,11 @@ class WorkflowRunner:
     ) -> str:
         # 1️⃣ manifest 优先
         if outputs:
-            manifest_item = self._manifest_character_by_role(outputs, "primary")
+            manifest_item = (
+                self._character_manifest_support.manifest_character_by_role(
+                    outputs, "primary"
+                )
+            )
             if manifest_item is not None:
                 display_value = str(manifest_item.get("display_name") or "").strip()
                 if display_value:
@@ -2209,7 +1924,11 @@ class WorkflowRunner:
         outputs: Optional[Dict[str, Any]] = None,
     ) -> str:
         if outputs:
-            manifest_item = self._manifest_character_by_role(outputs, "secondary")
+            manifest_item = (
+                self._character_manifest_support.manifest_character_by_role(
+                    outputs, "secondary"
+                )
+            )
             if manifest_item is not None:
                 display_value = str(manifest_item.get("display_name") or "").strip()
                 if display_value:
@@ -2279,7 +1998,9 @@ class WorkflowRunner:
         # ✅ 1) 优先走 character_manifest（primary 角色锚点）
         primary = None
         if outputs:
-            primary = self._manifest_character_by_role(outputs, "primary")
+            primary = self._character_manifest_support.manifest_character_by_role(
+                outputs, "primary"
+            )
 
         if isinstance(primary, dict):
             display = str(primary.get("display_name") or "").strip()
@@ -3485,7 +3206,9 @@ class WorkflowRunner:
                     f"整体气质{tone_label}，适合做成{audience_label}向内容。"
                 )
             title = f"{topic}的故事"
-        manifest_items = self._character_manifest_items(outputs)
+        manifest_items = self._character_manifest_support.character_manifest_items(
+            outputs
+        )
 
         return {
             "title": title,
